Match fclCombineList FUN_003e0950

+++ src/h_pad.c
@@ -417,7 +417,7 @@ void H_Pad_UpdateRumble(void)
             {
                 if (sRumblePhase == 0)
                 {
-                    u16 intensity;
+                    u32 intensity;
                     u8 intensityByte;
                     intensityByte = sRumbleIntensity.b;
                     intensity = intensityByte;
